Upload original buildable state
This is the first version which I was able to successfully compile on another platform. Has not been unit tested. Compiled using 'bsdmake' on mac OS X Ventura. Unfortunately, I also had to modify the Makefile to make this work. Hopefully, we could figure out a Makefile which could allow us to compile on various platforms without using different Makefiles. However, I do not have enough experience with the BSD make system to even know if that is possible, let alone how to implement it if it is. So, this is what we have for the moment...hopefully it is enough to cross-compile the source tree from OS X or Linux systems....